AG-7605 CR: Join crossLine.test.ts & crossLineEvent.test.ts
